requirements.txt
mysql-connector-python==8.0.28
numpy==1.22.3
opencv-contrib-python==4.5.5.64
Pillow==9.0.1
protobuf==3.19.4
